Transfer commit

+++ src/main/java/shader/Bufferable.java
@@ -2,5 +2,5 @@ package shader;
 
 public interface Bufferable {
     default int getSizePerObject() { return get().length; }
-    float[] get();
+    double[] get();
 }
